feat(search): implement jump search algorithm for singly linked list
Implemented the jump_list function in search_algos.h to perform a jump search in a sorted singly linked list of integers. The function takes a pointer to the head of the list, the number of nodes in the list, and the value to search for. It returns a pointer to the first node where the value is located or NULL if the value is not present or the head of the list is NULL. Additionally, the function prints the value every time it is compared in the list and uses the square root of the list size as the jump step.